+++ tools/tenscan/frontend/pages/tx/[hash].tsx
@@ -1,23 +1,71 @@
+import { fetchTransactionByHash } from "@/api/transactions";
 import Layout from "@/src/components/layouts/default-layout";
+import { TransactionDetailsComponent } from "@/src/components/modules/transactions/transaction-details";
 import EmptyState from "@/src/components/modules/common/empty-state";
 import { Button } from "@/src/components/ui/button";
+import {
+  Card,
+  CardHeader,
+  CardTitle,
+  CardContent,
+} from "@/src/components/ui/card";
+import { Skeleton } from "@/src/components/ui/skeleton";
+import { useQuery } from "@tanstack/react-query";
 import { useRouter } from "next/router";
-import React from "react";
 
-const TransactionDetails = () => {
-  const { push } = useRouter();
+export default function TransactionDetails() {
+  const router = useRouter();
+  const { hash } = router.query;
+
+  const { data, isLoading } = useQuery({
+    queryKey: ["transactionDetails", hash],
+    queryFn: () => fetchTransactionByHash(hash as string),
+  });
+
+  const transactionDetails = data?.item;
 
   return (
     <Layout>
-      <EmptyState
-        title="Transaction Details"
-        description="Coming soon..."
-        imageSrc="/assets/images/clock.png"
-        imageAlt="Clock"
-        action={<Button onClick={() => push("/")}>Go Home</Button>}
-      />
+      <Card className="col-span-3">
+        {isLoading ? (
+          <>
+            <Skeleton className="h-10 w-100" />
+            <Skeleton className="h-10 w-100" />
+            <Skeleton className="h-10 w-100" />
+            <Skeleton className="h-10 w-100" />
+            <Skeleton className="h-10 w-100" />
+            <Skeleton className="h-10 w-100" />
+          </>
+        ) : transactionDetails ? (
+          <>
+            <CardHeader>
+              <CardTitle>Transaction Details</CardTitle>
+            </CardHeader>
+            <CardContent>
+              <TransactionDetailsComponent
+                transactionDetails={transactionDetails}
+              />
+            </CardContent>
+          </>
+        ) : (
+          <EmptyState
+            title="Transaction not found"
+            description="The transaction you are looking for does not exist."
+            action={
+              <Button onClick={() => router.push("/transactions")}>
+                Go back
+              </Button>
+            }
+            className="p-8"
+          />
+        )}
+      </Card>
     </Layout>
   );
-};
+}
 
-export default TransactionDetails;
+export async function getServerSideProps(context: any) {
+  return {
+    props: {},
+  };
+}