+++ tools/tenscan/frontend/src/components/modules/dashboard/recent-batches.tsx
@@ -2,6 +2,7 @@ import TruncatedAddress from "../common/truncated-address";
 import { formatTimeAgo } from "@/src/lib/utils";
 import { Batch } from "@/src/types/interfaces/BatchInterfaces";
 import { Avatar, AvatarFallback } from "@/src/components/ui/avatar";
+import Link from "next/link";
 
 export function RecentBatches({ batches }: { batches: any }) {
   return (
@@ -13,14 +14,22 @@ export function RecentBatches({ batches }: { batches: any }) {
           </Avatar>
           <div className="ml-4 space-y-1">
             <p className="text-sm font-medium leading-none">
-              #{Number(batch?.number)}
+              <Link
+                href={`/batch/height/${batch?.height}`}
+                className="text-primary"
+              >
+                #{Number(batch?.height)}
+              </Link>
             </p>
             <p className="text-sm text-muted-foreground word-break-all">
-              {formatTimeAgo(batch?.timestamp)}
+              {formatTimeAgo(batch?.header?.timestamp)}
             </p>
           </div>
           <div className="ml-auto font-medium min-w-[140px]">
-            <TruncatedAddress address={batch?.hash} />
+            <TruncatedAddress
+              address={batch?.header?.hash}
+              link={`/batch/${batch?.header?.hash}`}
+            />
           </div>
         </div>
       ))}